+++ arch/parisc/include/asm/pgalloc.h
@@ -143,7 +143,7 @@ pte_alloc_one_kernel(struct mm_struct *mm, unsigned long addr)
 
 static inline void pte_free_kernel(struct mm_struct *mm, pte_t *pte)
 {
-	free_page((unsigned long)pte);
+	free_page(pte);
 }
 
 static inline void pte_free(struct mm_struct *mm, struct page *pte)
